requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
markdownify==0.11.6
aiohttp==3.9.1
asyncio==3.4.3
lxml==4.9.3
python-slugify==8.0.1
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1